+++ ica-rs/ica_typing.py
@@ -47,18 +47,18 @@ class IcaStatus:
         ...
 
 
-class ReplyMessage:
+class IcaReplyMessage:
     ...
 
 
-class SendMessage:
+class IcaSendMessage:
     @property
     def content(self) -> str:
         ...
     @content.setter
     def content(self, value: str) -> None:
         ...
-    def with_content(self, content: str) -> "SendMessage":
+    def with_content(self, content: str) -> "IcaSendMessage":
         """
         为了链式调用, 返回自身
         """
@@ -66,15 +66,15 @@ class SendMessage:
         return self
 
 
-class DeleteMessage:
+class IcaDeleteMessage:
     def __str__(self):
         ...
 
 
-class NewMessage:
-    def reply_with(self, message: str) -> SendMessage:
+class IcaNewMessage:
+    def reply_with(self, message: str) -> IcaSendMessage:
         ...
-    def as_deleted(self) -> DeleteMessage:
+    def as_deleted(self) -> IcaDeleteMessage:
         ...
     def __str__(self) -> str:
         ...
@@ -93,6 +93,15 @@ class NewMessage:
     @property
     def is_reply(self) -> bool:
         ...
+    @property
+    def is_room_msg(self) -> bool:
+        ...
+    @property
+    def is_chat_msg(self) -> bool:
+        ...
+    @property
+    def room_id(self) -> RoomId:
+        ...
 
 
 class IcaClient:
@@ -103,13 +112,13 @@ class IcaClient:
     #     (因为目前来说, rust调用 Python端没法启动一个异步运行时
     #     所以只能 tokio::task::block_in_place 转换成同步调用)
     #     """
-    def send_message(self, message: SendMessage) -> bool:
+    def send_message(self, message: IcaSendMessage) -> bool:
         ...
-    def send_and_warn(self, message: SendMessage) -> bool:
+    def send_and_warn(self, message: IcaSendMessage) -> bool:
         """发送消息, 并在日志中输出警告信息"""
         self.warn(message.content)
         return self.send_message(message)
-    def delete_message(self, message: DeleteMessage) -> bool:
+    def delete_message(self, message: IcaDeleteMessage) -> bool:
         ...
     
     @property
@@ -144,11 +153,11 @@ on_load = Callable[[IcaClient], None]
 # def on_load(client: IcaClient) -> None:
 #     ...
 
-on_message = Callable[[NewMessage, IcaClient], None]
+on_ica_message = Callable[[IcaNewMessage, IcaClient], None]
 # def on_message(msg: NewMessage, client: IcaClient) -> None:
 #     ...
 
-on_delete_message = Callable[[MessageId, IcaClient], None]
+on_ica_delete_message = Callable[[MessageId, IcaClient], None]
 # def on_delete_message(msg_id: MessageId, client: IcaClient) -> None:
 #     ...